requirements.txt
fastapi
uvicorn
pandas
numpy
xgboost
ccxt
pyyaml
scikit-learn
APScheduler
sqlalchemy
pymysql